requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==22.0.0
eventlet==0.36.1
psycopg2-binary==2.9.9
